==> 00-author.Rmd <==
# About the Authors {-}

**Sarah Lin** manages the Enterprise Information Management team at Posit, PBC.  A graduate of the University of Illinois iSchool, Sarah worked as a technical services librarian in many different library types (academic, special, medical & law) before moving into corporate librarianship and information management.  Her professional interests include findability and metadata, and how metadata enables findability.  Sarah is a certified Carpentries instructor and received her undergraduate degree in African/African-American Studies & Anthropology from the University of Chicago.  She didn't know anything about coding in R before joining Posit.

**Dorris Scott** is currently the Academic Director of Data Studies at Washington University in St. Louis - University College. As Academic Director of Data Studies, Dorris develops new curriculum, certificate and degree programs related to data analytics, data science, and geographic information systems (GIS) along with forges local and regional partnerships in alignment with University College’s strategic vision. Previously, Dorris was was Geographic Information Systems (GIS) Librarian and Social Science Data Curator at University Libraries at Washington University and provided consultation on projects that use geospatial data along with providing training in various GIS software, programming applications of geospatial data, and data management. As part of this position, Dorris also served as a liaison between Washington University Libraries and social science departments assisting faculty with their data needs such as data management and data curation. Dorris received a PhD in Geography from the University of Georgia, with a specialization in GIS applications for public health.
